+++ src/modules/patient/controllers/test.controller.ts
@@ -1,18 +1,18 @@
 import {
-  Controller,
-  Get,
-  Post,
-  Body,
-  Patch,
-  Param,
-  Delete,
-  NotFoundException,
   BadRequestException,
+  Body,
+  Controller,
+  Delete,
+  Get,
+  NotFoundException,
+  Param,
+  Patch,
+  Post,
   Query,
 } from '@nestjs/common';
 import { TestService } from '../services';
 import { CreateTestDto, TestPaginationDto, UpdateTestDto } from '../dto';
-import { test as TestModel } from '@prisma/client';
+import { PatientPosition } from '@prisma/client';
 
 @Controller('test')
 export class TestController {
@@ -29,25 +29,33 @@ export class TestController {
     return this.modelService.list(params);
   }
 
+  @Get('positions')
+  getListPosition() {
+    return PatientPosition;
+  }
+
   @Get(':id')
   get(@Param('id') id: string) {
+    if (!+id) throw new BadRequestException('The id must be a number');
     return this.getInstanceOr404(+id);
   }
 
   @Post()
   async create(@Body() createTestDto: CreateTestDto) {
-    const testModel: TestModel = await this.modelService.create(createTestDto);
+    const testModel = await this.modelService.create(createTestDto);
     if (!testModel) throw new BadRequestException('Invalid test!');
     return testModel;
   }
 
   @Patch(':id')
   async update(@Param('id') id: string, @Body() updateTestDto: UpdateTestDto) {
+    if (!+id) throw new BadRequestException('The id must be a number');
     return this.modelService.update(+id, updateTestDto);
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string) {
+    if (!+id) throw new BadRequestException('The id must be a number');
     return this.modelService.delete(+id);
   }
 }
